Split off sample gcal into method calls.
